requirements.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
protobuf==3.20.1
pycodestyle==2.10.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
tenacity==8.1.0
tomli==2.0.1
Werkzeug==2.2.2
